Refactor how versioning information is indicated in the build and in the source.
